Begin tidying up sequencer classes
