+++ Source/BKEditorUtilities/Private/SequenceAssetEditor/UnDAWSequenceEditorToolkit.cpp
@@ -271,19 +271,26 @@ void FUnDAWSequenceEditorToolkit::RegisterTabSpawners(const TSharedRef<class FTa
 	FPropertyEditorModule& PropertyEditorModule = FModuleManager::GetModuleChecked<FPropertyEditorModule>("PropertyEditor");
 	FDetailsViewArgs DetailsViewArgs;
 	DetailsViewArgs.NameAreaSettings = FDetailsViewArgs::HideNameArea;
-	TSharedRef<IDetailsView> DetailsView = PropertyEditorModule.CreateDetailView(DetailsViewArgs);
-	DetailsView->SetObjects(TArray<UObject*>{ SequenceData });
+	DetailsViewArgs.bAllowMultipleTopLevelObjects = true;
+	DetailsViewArgs.bAllowSearch = true;
+	DetailsViewArgs.bShowSectionSelector = true;
+	DetailsViewArgs.bAllowFavoriteSystem = true;
+	DetailsViewArgs.bShowOptions = true;
+	DetailsView = PropertyEditorModule.CreateDetailView(DetailsViewArgs).ToSharedPtr();
+	//DetailsView->
+	//DetailsView->SetObjects(TArray<UObject*>{ SequenceData });
 
 	// FDetailsViewArgs NodeDetailsViewArgs;
    //  NodeDetailsViewArgs.NameAreaSettings = FDetailsViewArgs::ObjectsUseNameArea;
 
 	 //NodeDetailsView = PropertyEditorModule.CreateDetailView(NodeDetailsViewArgs);
 
-	InTabManager->RegisterTabSpawner("DAWSequenceDetailsTab", FOnSpawnTab::CreateLambda([=](const FSpawnTabArgs&)
+	InTabManager->RegisterTabSpawner("DAWSequenceDetailsTab", FOnSpawnTab::CreateLambda([this](const FSpawnTabArgs&)
 		{
 			return SNew(SDockTab)
 				[
-					DetailsView
+					DetailsView.ToSharedRef()
+
 
 				];
 		}))
@@ -331,6 +338,8 @@ TSharedRef<SButton> FUnDAWSequenceEditorToolkit::GetConfiguredTransportButton(EB
 void FUnDAWSequenceEditorToolkit::OnSelectionChanged(const TSet<UObject*>& SelectedNodes)
 {
 	UM2SoundGraph* Graph = Cast<UM2SoundGraph>(SequenceData->M2SoundGraph);
+
+	DetailsView->SetObjects(SelectedNodes.Array());
 
 	Graph->SelectedNodes.Empty();
 	Graph->SelectedVertices.Empty();
